Moved out aws responses from db.py into aws.py
